+++ data-error/Cargo.toml
@@ -1,8 +1,7 @@
 [package]
-name = "fs-utils"
+name = "data-error"
 version = "0.1.0"
 edition = "2021"
-
 
 [dependencies]
 thiserror = "1"
@@ -10,7 +9,3 @@ reqwest = "0.11.11"
 serde_json = "1.0.82"
 anyhow = "1"
 url = { version = "2.2.2", features = ["serde"] }
-
-[dev-dependencies]
-rstest = "0.18"
-tempdir = "0.3"
