+++ data-json/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "data-json"
+version = "0.1.0"
+edition = "2021"
+
+
+[dependencies]
+serde_json = "1.0.82"
+
+[dev-dependencies]
+rstest = "0.18"
