+++ data-pdf/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "data-pdf"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+name = "data_pdf"
+crate-type = ["rlib"]
+bench = false
+
+[dependencies]
+once_cell = "1.16.0"
+image = "0.24.2"
+pdfium-render = { git = "https://github.com/ajrcarey/pdfium-render", rev = "d2559c1", features = [
+    "thread_safe",
+    "sync",
+] }
+
+[dev-dependencies]
+tempdir = "0.3.7"
+
+[build-dependencies]
+flate2 = "1.0.24"
+fs_extra = "1.2.0"
+tar = "0.4.38"
+target-lexicon = "0.12.4"
+ureq = "2.4.0"
+ring = "=0.17.5"
