+++ data-resource/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-fs-utils = { path = "../fs-utils" }
+data-error = { path = "../data-error" }
 fs-atomic-versions = { path = "../fs-atomic-versions" }
 
 log = { version = "0.4.17", features = ["release_max_level_off"] }
