+++ fs-atomic-light/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "fs-atomic-light"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+name = "fs_atomic_light"
+crate-type = ["rlib"]
+bench = false
+
+[dependencies]
+data-error = { path = "../data-error" }
