+++ fs-index/Cargo.toml
@@ -9,8 +9,9 @@ crate-type = ["rlib"]
 bench = false
 
 [dependencies]
-fs-utils = { path = "../fs-utils" }
+data-error = { path = "../data-error" }
 fs-atomic-versions = { path = "../fs-atomic-versions" }
+fs-storage = { path = "../fs-storage" }
 data-resource = { path = "../data-resource" }
 
 log = { version = "0.4.17", features = ["release_max_level_off"] }
@@ -19,20 +20,7 @@ anyhow = "1.0.58"
 lazy_static = "1.4.0"
 canonical-path = "2.0.2"
 pathdiff = "0.2.1"
-serde_json = "1.0.82"
-serde = { version = "1.0.138", features = ["derive"] }
-url = { version = "2.2.2", features = ["serde"] }
-reqwest = "0.11.11"
-scraper = "0.13.0"
-tokio = { version = "1", features = ["full"] }
 itertools = "0.10.5"
-once_cell = "1.16.0"
-image = "0.24.2"
-pdfium-render = { git = "https://github.com/ajrcarey/pdfium-render", rev = "d2559c1", features = [
-    "thread_safe",
-    "sync",
-] }
-
 
 [dev-dependencies]
 uuid = { version = "1.6.1", features = ["v4"] }
@@ -42,14 +30,6 @@ rstest = '0.18.2'
 criterion = { version = "0.5", features = ["html_reports"] }
 pprof = { version = "0.13", features = ["criterion", "flamegraph"] }
 rand = "0.8"
-
-[build-dependencies]
-flate2 = "1.0.24"
-fs_extra = "1.2.0"
-tar = "0.4.38"
-target-lexicon = "0.12.4"
-ureq = "2.4.0"
-ring = "=0.17.5"
 
 [[bench]]
 name = "index_build_benchmark"
