--- fs-index/src/storage/mod.rs
@@ -1,2 +1,0 @@
-pub mod meta;
-pub mod prop;
