--- fs-utils/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod errors;
-pub mod json;
